requirements.txt
# Core AI model providers
openai>=1.13.3
anthropic>=0.25.8
google-generativeai>=0.6.0

# Helpful utilities
tqdm>=4.66.1
python-dotenv>=1.0.1

# Data handling (CSV/JSON/metadata)
pandas>=2.2.1
discord.py>=2.0.0